refactor: rename style-classes
