+++ src/components/buttons/styles/TextLinkButtonStyles.js
@@ -1,39 +1,33 @@
 import { StyleSheet } from "react-native";
-import theme from "../../theme";
+import theme from "../../../theme";
 
 const styles = StyleSheet.create({
     container: {
-        display: 'flex',
+        flex: 1,
         flexDirection: 'row',
-        justifyContent: 'left',
+        alignSelf: 'flex-start',
         alignItems: 'center',
         borderStyle: 'solid',
         borderWidth: '3px',
         borderColor: theme.PRIMARY_COLOR,
         backgroundColor: 'black',
-        padding: '25px',
+        padding: '10px',
+        margin: '5px',
+        marginTop: '15px',
         borderRadius: '15px',
         textDecorationLine: 'none',
-        marginBottom: '10px',
-        height: '50px',
-        width: '250px'
     },
     containerHover: {
         backgroundColor: theme.PRIMARY_COLOR,
     },
     text: {
-        marginLeft: '10px',
-        justifyContent: 'center',
+        flex: 1,
         textAlignVertical: 'center',
         textAlign: 'center',
         color: theme.PRIMARY_COLOR,
     },
     textHover: {
         color: theme.SECONDARY_COLOR,
-    },
-    icon: {
-        width: '50px',
-        height: '50px',
     },
 });
