+++ src/components/containers/domains/social-media/SocialMediaLinks.js
@@ -1,6 +1,6 @@
 import { Text, View } from 'react-native';
 import SocialMedia from './SocialMedia';
-import styles from './styles/style-SocialMediaLinks';
+import styles from './styles/SocialMediaLinksStyles';
 import theme from '../../../../theme';
 import SocialMediaButton from './SocialMediaButton';
